# docs/api/federation/openapi.yaml
openapi: 3.0.3
info:
  title: NexixAI AgentOS — Federation API
  version: 1.0.2
  description: 'Federation connects AgentOS nodes across environments (peer info/capabilities,
    run forwarding, and event backhaul). Contract authority: docs/product/agentos-prs/v1.02-schemas-appendix.md'
servers:
- url: /
tags:
- name: Health
- name: Peer
- name: Forwarding
- name: Events
security:
- bearerAuth: []
- apiKeyAuth: []
paths:
  /v1/federation/health:
    get:
      tags:
      - Health
      summary: Federation health check
      operationId: federationHealth
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/federation/peer:
    get:
      tags:
      - Peer
      summary: Get local peer info
      operationId: getPeerInfo
      parameters:
      - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerInfoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/federation/peer/capabilities:
    get:
      tags:
      - Peer
      summary: Get local peer capabilities
      operationId: getPeerCapabilities
      parameters:
      - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerCapabilitiesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/federation/runs:forward:
    post:
      tags:
      - Forwarding
      summary: Forward a run to a remote peer
      operationId: forwardRun
      parameters:
      - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationForwardRunRequest'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationForwardRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/federation/runs/{run_id}/events:
    get:
      tags:
      - Events
      summary: Stream events for a run (SSE)
      operationId: streamFederatedRunEvents
      parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/XCorrelationId'
      - name: after_sequence
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: SSE stream
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/federation/events:ingest:
    post:
      tags:
      - Events
      summary: Ingest backhauled events (push mode)
      operationId: ingestFederatedEvents
      parameters:
      - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationEventIngestRequest'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationEventIngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  headers:
    XRequestId:
      description: Request id assigned by the service (or propagated).
      schema:
        type: string
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    RateLimited:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/ErrorResponse
  schemas:
    HealthResponse:
      type: object
      required:
      - status
      - service
      - version
      properties:
        status:
          type: string
          enum:
          - ok
        service:
          type: string
          enum:
          - federation
        version:
          type: string
    PeerInfoResponse:
      type: object
      required:
      - peer
      properties:
        peer:
          $ref: '#/components/schemas/PeerInfo'
    PeerInfo:
      type: object
      required:
      - stack_id
      - environment
      - region
      - api_versions
      - endpoints
      - build
      properties:
        stack_id:
          type: string
        environment:
          type: string
        region:
          type: string
        api_versions:
          type: array
          items:
            type: string
        endpoints:
          $ref: '#/components/schemas/PeerEndpoints'
        build:
          $ref: '#/components/schemas/PeerBuild'
    PeerEndpoints:
      type: object
      required:
      - agent-orchestrator_base_url
      - model-policy_base_url
      properties:
        agent-orchestrator_base_url:
          type: string
        model-policy_base_url:
          type: string
    PeerBuild:
      type: object
      required:
      - version
      - git_sha
      - timestamp
      properties:
        version:
          type: string
        git_sha:
          type: string
        timestamp:
          type: string
          format: date-time
    PeerCapabilitiesResponse:
      type: object
      required:
      - capabilities
      properties:
        capabilities:
          $ref: '#/components/schemas/PeerCapabilities'
    PeerCapabilities:
      type: object
      required:
      - supports
      - limits
      - models
      properties:
        supports:
          $ref: '#/components/schemas/CapabilitySupports'
        limits:
          $ref: '#/components/schemas/CapabilityLimits'
        models:
          $ref: '#/components/schemas/CapabilityModels'
    CapabilitySupports:
      type: object
      required:
      - runs_forward
      - events_stream
      - tenant_isolation
      - tool_hosting
      properties:
        runs_forward:
          type: boolean
        events_stream:
          type: boolean
        tenant_isolation:
          type: boolean
        tool_hosting:
          type: boolean
    CapabilityLimits:
      type: object
      required:
      - max_concurrent_runs
      - max_event_streams
      properties:
        max_concurrent_runs:
          type: integer
          minimum: 0
        max_event_streams:
          type: integer
          minimum: 0
    CapabilityModels:
      type: object
      required:
      - tiers
      - backends
      - max_context_tokens
      properties:
        tiers:
          type: array
          items:
            type: string
        backends:
          type: array
          items:
            type: string
        max_context_tokens:
          type: integer
          minimum: 0
    FederationForwardRunRequest:
      type: object
      required:
      - forward
      properties:
        forward:
          $ref: '#/components/schemas/ForwardPayload'
    ForwardPayload:
      type: object
      required:
      - target_selector
      - auth
      - run_request
      - traceparent
      properties:
        target_selector:
          $ref: '#/components/schemas/TargetSelector'
        auth:
          $ref: ../model-policy/openapi.yaml#/components/schemas/AuthContext
        run_request:
          $ref: '#/components/schemas/RunForwardRequest'
        traceparent:
          type: string
    TargetSelector:
      type: object
      required:
      - stack_id
      - region
      - required_capabilities
      - preferred_model_tier
      properties:
        stack_id:
          type: string
        region:
          type: string
        required_capabilities:
          type: array
          items:
            type: string
        preferred_model_tier:
          type: string
          enum:
          - fast
          - balanced
          - premium
    RunForwardRequest:
      type: object
      required:
      - agent_id
      - input
      - context
      - tooling
      - run_options
      - idempotency_key
      properties:
        agent_id:
          type: string
        input:
          $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/RunInput
        context:
          $ref: '#/components/schemas/RunForwardContext'
        tooling:
          $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/Tooling
        run_options:
          $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/RunOptions
        idempotency_key:
          type: string
          maxLength: 256
    RunForwardContext:
      type: object
      required:
      - locale
      - timezone
      properties:
        locale:
          type: string
        timezone:
          type: string
        channel:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        metadata:
          type: object
          additionalProperties: true
    FederationForwardRunResponse:
      type: object
      required:
      - forwarded
      - correlation_id
      properties:
        forwarded:
          $ref: '#/components/schemas/ForwardedResult'
        correlation_id:
          type: string
    ForwardedResult:
      type: object
      required:
      - tenant_id
      - remote_stack_id
      - remote_run_id
      - remote_events_url
      - status
      properties:
        tenant_id:
          type: string
        remote_stack_id:
          type: string
        remote_run_id:
          type: string
        remote_events_url:
          type: string
        status:
          type: string
          enum:
          - queued
          - running
          - completed
          - failed
    FederationEventIngestRequest:
      type: object
      required:
      - peer
      - events
      properties:
        peer:
          $ref: '#/components/schemas/PeerInfo'
        events:
          type: array
          items:
            $ref: ../agent-orchestrator/openapi.yaml#/components/schemas/EventEnvelope
        traceparent:
          type: string
    FederationEventIngestResponse:
      type: object
      required:
      - accepted
      - rejected
      - correlation_id
      properties:
        accepted:
          type: integer
          minimum: 0
        rejected:
          type: integer
          minimum: 0
        correlation_id:
          type: string